requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.23.0
rapidocr-onnxruntime
onnxruntime
Pillow>=10.0.0
numpy>=1.24.0
